Subject: Edited README. Added the router, finally. Landing page added to and moved to completion.
Context: Landing page was moved into the root of the router so that the rest of the files could be served from there. Additionally the ladning page had additonal elements and information added to it.
File structure edited with additons for more components and things Like testimonials and stuff, but nothing's in there so that does not make the subject line. Whee I think we're done for thsi commit!